T #96 Cache service rename
