+++ Huppy/Huppy.Core/Attributes/RequireDevAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using Discord;
 using Discord.Interactions;
-using Huppy.Core.Services.HuppyCache;
+using Huppy.Core.Services.HuppyCacheStorage;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Huppy.Core.Attributes;
@@ -8,7 +8,7 @@ public class RequireDevAttribute : PreconditionAttribute
 {
     public override Task<PreconditionResult> CheckRequirementsAsync(IInteractionContext context, ICommandInfo commandInfo, IServiceProvider services)
     {
-        var cacheService = services.GetRequiredService<CacheService>();
+        var cacheService = services.GetRequiredService<CacheStorageService>();
 
         bool isUserDev = cacheService.DeveloperIds.Contains(context.User.Id);
